+++ packages/firebase/index.d.ts
@@ -5486,13 +5486,13 @@ declare namespace firebase.storage {
     /**
      * List all items (files) and prefixes (folders) under this storage reference.
      *
-     * This is a helper method for calling list() repeatedly until there are
+     * This is a helper method for calling `list()` repeatedly until there are
      * no more results. The default pagination size is 1000.
      *
      * Note: The results may not be consistent if objects are changed while this
      * operation is running.
      *
-     * Warning: listAll may potentially consume too many resources if there are
+     * Warning: `listAll` may potentially consume too many resources if there are
      * too many results.
      *
      * @return A Promise that resolves with all the items and prefixes under
@@ -5513,9 +5513,9 @@ declare namespace firebase.storage {
      * To adhere to Firebase Rules's Semantics, Firebase Storage does not
      * support objects whose paths end with "/" or contain two consecutive
      * "/"s. Firebase Storage List API will filter these unsupported objects.
-     * list() may fail if there are too many unsupported objects in the bucket.
+     * `list()` may fail if there are too many unsupported objects in the bucket.
      *
-     * @param options See ListOptions for details.
+     * @param options See `ListOptions` for details.
      * @return A Promise that resolves with the items and prefixes.
      *      `prefixes` contains references to sub-folders and `items`
      *      contains references to objects in this folder. `nextPageToken`
@@ -5549,7 +5549,7 @@ declare namespace firebase.storage {
   }
 
   /**
-   * The options list() accepts.
+   * The options `list()` accepts.
    */
   interface ListOptions {
     /**
@@ -5558,7 +5558,7 @@ declare namespace firebase.storage {
      */
     maxResults?: number | null;
     /**
-     * The `nextPageToken` from a previous call to list(). If provided,
+     * The `nextPageToken` from a previous call to `list()`. If provided,
      * listing is resumed from the previous position.
      */
     pageToken?: string | null;
@@ -6086,7 +6086,8 @@ declare namespace firebase.firestore {
     /**
      * Attempts to enable persistent storage, if possible.
      *
-     * Must be called before any other methods (other than settings()).
+     * Must be called before any other methods (other than settings() and
+     * clearPersistence()).
      *
      * If this fails, enablePersistence() will reject the promise it returns.
      * Note that even after this failure, the firestore instance will remain
@@ -6177,6 +6178,28 @@ declare namespace firebase.firestore {
      * instance.
      */
     app: firebase.app.App;
+
+    /**
+     * Clears the persistent storage. This includes pending writes and cached
+     * documents.
+     *
+     * Must be called while the firestore instance is not started (after the app is
+     * shutdown or when the app is first initialized). On startup, this method
+     * must be called before other methods (other than settings()). If the
+     * firestore instance is still running, the promise will be rejected with
+     * the error code of `failed-precondition`.
+     *
+     * Note: clearPersistence() is primarily intended to help write reliable
+     * tests that use Firestore. It uses the most efficient mechanism possible
+     * for dropping existing data but does not attempt to securely overwrite or
+     * otherwise make cached data unrecoverable. For applications that are
+     * sensitive to the disclosure of cache data in between user sessions we
+     * strongly recommend not to enable persistence in the first place.
+     *
+     * @return A promise that is resolved once the persistent storage has been
+     * cleared. Otherwise, the promise is rejected with an error.
+     */
+    clearPersistence(): Promise<void>;
 
     /**
      * Re-enables use of the network for this Firestore instance after a prior